Created a javafx project and imported previous stranded into project, AB#272
