+++ strandedfx/resources/hidden.json
