+++ strandedfx/resources/items.json
