+++ strandedfx/resources/planet1.json
