+++ strandedfx/src/com/game/conditions/Lose.java
