+++ strandedfx/src/com/game/conditions/Travel.java
