+++ strandedfx/src/com/game/textparser/Directions.java
